% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent.R
\name{concurrent}
\alias{concurrent}
\title{Concurrent Model for Functional Regression}
\usage{
concurrent()
}
\value{
A training and a prediction function.
}
\description{
It is a concurrent model, which may be fed to \code{\link{conformal.fun.split}}.
}
\details{
For more details about the structure of the inputs go to split.R
}
